// client/types.gen.ts
export enum LogType {
    HOST_CHANGED = 0,
    TURN_START = 1,
    TIMEOUT = 2,
    CHECK = 3,
    FOLD = 4,
    CALL = 5,
    RAISE = 6,
    ALL_IN = 7,
    LEAVE = 8,
    WIN = 9,
    GAME_CREATED = 10,
    FLOP = 11,
    TURN = 12,
    RIVER = 13,
    BUY_IN = 14,
    TOP_UP = 15,
    POST_SB = 16,
    POST_BB = 17,
    HAND_START = 18,
    HAND_END = 19,
    RENAME = 20,
    UNCALLED_BET = 21,
    SHOW_CARD = 22,
    SHOW_CARDS = 23,
    STRADDLE = 24,
    FLOP_RUN_IT_TWICE = 25,
    TURN_RUN_IT_TWICE = 26,
    RIVER_RUN_IT_TWICE = 27,
    BOARD_WIN = 28,
    BET = 29,
    PLAYERS = 30,
    BOMB_POT = 31,
    POST_BOMB_POT = 32,
    POST_ANTE = 33,
    RUN_WIN = 34,
}

export enum HandName {
    NONE = 0,
    LAST_PLAYER_STANDING = 1,
    HIGH_CARD = 2,
    PAIR = 3,
    TWO_PAIR = 4,
    THREE_OF_A_KIND = 5,
    STRAIGHT = 6,
    FLUSH = 7,
    FULL_HOUSE = 8,
    FOUR_OF_A_KIND = 9,
    STRAIGHT_FLUSH = 10,
    ROYAL_FLUSH = 11,
}